rename ECSqlReader to QueryReader
