+++ core/common/src/QueryReader.ts
@@ -76,7 +76,7 @@ export interface QueryStats {
 }
 
 /** @beta */
-export class ECSqlReader {
+export class QueryReader {
   private static readonly _maxRetryCount = 10;
   private _localRows: any[] = [];
   private _localOffset: number = 0;
@@ -88,8 +88,8 @@ export class ECSqlReader {
   private _param = new QueryBinder().serialize();
   private _lockArgs: boolean = false;
   private _stats = { backendCpuTime: 0, backendTotalTime: 0, backendMemUsed: 0, backendRowsReturned: 0, totalTime: 0, retryCount: 0 };
-  private _rowProxy = new Proxy<ECSqlReader>(this, {
-    get: (target: ECSqlReader, key: string | Symbol) => {
+  private _rowProxy = new Proxy<QueryReader>(this, {
+    get: (target: QueryReader, key: string | Symbol) => {
       if (typeof key === "string") {
         const idx = Number.parseInt(key, 10);
         if (!Number.isNaN(idx)) {
@@ -111,10 +111,10 @@ export class ECSqlReader {
       }
       return undefined;
     },
-    has: (target: ECSqlReader, p: string | symbol) => {
+    has: (target: QueryReader, p: string | symbol) => {
       return !target._props.findByNoCase(p as string);
     },
-    ownKeys: (target: ECSqlReader) => {
+    ownKeys: (target: QueryReader) => {
       const keys = [];
       for (const prop of target._props) {
         keys.push(prop.name);
@@ -195,7 +195,7 @@ export class ECSqlReader {
     if (this._globalCount === 0) {
       return [];
     }
-    const valueFormat = this._options.rowFormat === QueryRowFormat.UseJsPropertyNames? DbValueFormat.JsNames :DbValueFormat.ECSqlNames;
+    const valueFormat = this._options.rowFormat === QueryRowFormat.UseJsPropertyNames ? DbValueFormat.JsNames : DbValueFormat.ECSqlNames;
     const request: DbQueryRequest = {
       ... this._options,
       kind: DbRequestKind.ECSql,
@@ -211,7 +211,7 @@ export class ECSqlReader {
       this._props = new PropertyMetaDataMap(resp.meta);
     }
     for (const row of resp.data) {
-      ECSqlReader.replaceBase64WithUint8Array(row);
+      QueryReader.replaceBase64WithUint8Array(row);
     }
     return resp.data;
   }
@@ -229,7 +229,7 @@ export class ECSqlReader {
       this.stats.totalTime += (Date.now() - startTime);
       return rs;
     };
-    let retry = ECSqlReader._maxRetryCount;
+    let retry = QueryReader._maxRetryCount;
     let resp = await execQuery(request);
     DbQueryError.throwIfError(resp, request);
     while (--retry > 0 && needRetry(resp)) {
