+++ core/common/src/core-common.ts
@@ -82,7 +82,7 @@ export * from "./ModelProps";
 export * from "./NativeAppProps";
 export * from "./OctEncodedNormal";
 export * from "./ConcurrentQuery";
-export * from "./ECSqlReader";
+export * from "./QueryReader";
 export * from "./PlanarClipMask";
 export * from "./ModelGeometryChanges";
 export * from "./PlanProjectionSettings";
